Creating the resource archive is extracted to framework-specific spots
